Initial introduction of tower-service-util
