+++ tower-service-util/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+
+name = "tower-service-util"
+# When releasing to crates.io:
+# - Update html_root_url.
+# - Update CHANGELOG.md.
+# - Update documentation URL
+# - Create "v0.x.y" git tag.
+version = "0.1.0"
+authors = ["Carl Lerche <me@carllerche.com>"]
+license = "MIT"
+readme = "README.md"
+repository = "https://github.com/tower-rs/tower"
+homepage = "https://github.com/tower-rs/tower"
+documentation = "https://docs.rs/tokio-service-util/0.1.0"
+description = """
+Utilities for working with tower-service.
+"""
+categories = ["asynchronous", "network-programming"]
+
+[features]
+io = ["tokio-io"]
+
+[dependencies]
+futures = "0.1.23"
+tower-service = "0.2.0"
+tokio-io = { version = "0.1.12", optional = true }
