+++ tower-service-util/README.md
@@ -1,0 +1,13 @@
+# Tower Service Util
+
+Utilities for working with `tower-service`.
+
+## License
+
+This project is licensed under the [MIT license](LICENSE).
+
+### Contribution
+
+Unless you explicitly state otherwise, any contribution intentionally submitted
+for inclusion in Tower by you, shall be licensed as MIT, without any additional
+terms or conditions.
