+++ tower-service-util/src/sealed.rs
@@ -1,0 +1,1 @@
+pub trait Sealed<T> {}
