+++ tower-util/Cargo.toml
@@ -7,7 +7,7 @@ publish = false
 [dependencies]
 futures = "0.1"
 tower-service = "0.2.0"
-tokio-io = "0.1"
+tower-service-util = { version = "0.1.0", path = "../tower-service-util", features = ["io" ] }
 
 [dev-dependencies]
 tokio-mock-task = "0.1"
